+++ src/main/java/com/my/challenger/dto/quiz/AiValidationResult.java
@@ -5,13 +5,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class AiValidationResult {
     private boolean equivalent;
-    private double confidence;
+    private BigDecimal confidence;
     private String explanation;
     private boolean aiUsed;
     private boolean fallbackUsed;
